+++ src/PSI1620M_GiuliaTorres_2220119_PROJETO/PSI1620M_GiuliaTorres_2220119_PROJETO/Cprodutos.cs
@@ -12,6 +12,6 @@ namespace PSI1620M_GiuliaTorres_2220119_PROJETO
         public string ProdutoNome { get; set; }
         public int ProdutoIdVendedor { get; set; }
         public string ProdutoDescricao { get; set; }
-        public float ProdutoValor { get; set; }
+        public decimal ProdutoValor { get; set; }
     }
 }
